libxb: Add a method of sending AT commands from API mode.
This patch creates a new function for XbeeModule; xbee_at_command().
This function will add a frame to the XBee module's outgoing frame
queue that contains an AT command.

In order to request a response from the AT command's execution on
the XBee module, xbee_at_command puts a non-zero frame ID in the
outgoing frame.  An array of frame ID callback functions is added
to the XBeeModule member variables, which will get called when the
response is received.

The frame ID code could be extended in the future to receive the TX
status responses.