made all the changes in robs email:
Motor board polarity:  This is indicated in the board outlines in
subversion at: /media/board outlines

PWM board polarities: The spades for the power board are indicated on
the board outline.  I've just updated the board outline to include the
polarity of the servo connections too.

Reading through the document:

 "Be sure to line up the connector carefully, since it can only fit in
one orientation" - this contradicts itself.  It only fits in one way, so
one doesn't have to be as careful!

"Batter Charger" <- missing 'y'

I've just modified the board outline diagram for the power board to have
the labels that are on the PCB silk screen - I put the result
in /assemblies/assembly-guide/images.

"This will allow power to the logic to be disconnect quickly if
necessary." -> should probably be something along the lines of "This is
the main power switch for the robot.".

I felt that the charge-run switch description needed better description
- so I created a diagram, which is also
in /assemblies/assembly-guide/images.

Still some confusion about the motor board and pwm board naming:

 * PWM Board: this connects to servos.  You have referred to it as the
Servo board in places.
 * Motor board: This connectors motors.  You have referred to it as the
PWM board in places.

"It is advisable to use the mains adaptor for charging the battery only,
i.e. not for powering the robot during development." - this is not true.
The reason we have the charge-run switch is so that you can develop the
robot when the battery's flat.

"If your design requires additional motors, e.g. for moving armatures,
the power for these should be sourced from the power terminals of the
Motor Controller" -> perhaps it would be better to say the motor rail
off the power board or something.

"Nothing should be connected directly to the battery. This includes
additional motors and circuitry. (see previous paragraph for
instructions on connecting additional motors)" -> except for the power
board!

Competition mode is enabled/disabled by flicking one of the DIP
switches.  It's hard-coded into the power board, so they can't just
choose a switch for it.  (I've asked Tom if he can remember which switch
it is - should get an answer soon).