roboide: Allow switching between teams.
This patch updates ProjPage to have a destroy() member function.  The
most relevant part of this function to this commit is the bit where it
removes the signal handlers that it has attached to static bits of the
document.

If this clean-up isn't done, then when we call show() again, new
handlers are attached to the events -- causing things to be done N
times -- where N is the number of times we've changed team.