roboide: Make a class for the EditPage.
There is one instance of the EditPage.  It is in charge of the
components on the edit tabs.

This patch adds a new signal to the TabBar: "onswitch", which is
generated when the tab is changed.  The recipient of this event is
sent the tab that the switch is from and the one that it's to.  This
allows the EditPage to work out when it should be showing/hiding the
editarea -- allowing faster tab transitions.

Another big difference between this and the previous implementation is
that there is only one editarea instance.