added full revert support. fixed bug in radio buttons on log page. Now there is a warning message if you attempt to revert to the current version.
i have added the log history to the edit tabs now. You can choose to open any of the last ten versions of a file from the drop down box or you can view the full log with all revisions in a new tab by selecting the 'view full history' item at the bottom of the list. 

A nice feature is that on hover over the different versions you can see the corresponding commit messages