roboide: A very basic tab for showing code errors.
Code checking is going to be very simple to start with.  I'm simply
going to pump out the output of pychecker straight to the user.

This patch implements the ErrorsTab class, which will show the output
to the user.  It also implements a placeholder handler in the backend
that always returns the same thing at the moment.