Removed the directory 'chris' which was a previous attempt at rewriting the firmware for the pwm board.
Re-written the firmware for the new pwm board. I have ported the code from the previous v1 firmware but have had to change most of the initialisation routine for the new version. At present i have only ported the servo controlling code, i.e. not the I2C but this should be very simple to do. 

The routine loaded at present runs the servoSweep() function to sweep all of the servo outputs simultaneously. The routine has been running successfully on my desk for oven an hour now with two servos connected. 

Yet to do:
> Fix the short on port pin 2.1 - the status LED is never illuminated
> Re-enable the code to check for the presence of VCC
> Port I2C routines from motor firmware to control the servos.

msp430-gdb* will be the death of me!