flashb: Support the two different command formats that we have.
We have two varieties of I2C command formats.  One uses the SMBus
byte and word functions for bytes and words, and uses write() to write
fixed sized blocks of data containing no length byte.  The other
exclusively uses read() and write() to read/write blocks of data.  All
these blocks contain the length.

The motor, jointio and pwm are in the former scheme, whilst the power
board is in the latter.