flash430: Store the firmware version in the MSP430 flash.
This patch makes flash430, the I2C bootloader, store a 2-byte firmware
revision number at the beginning of flash.  It also connects up the
I2C handler to this value.

The value is stored in flash using by adding a new section, ".fwver",
which the linker script places at the beginning of flash.