roboide: Don't resize the commit message textbox.
This patch gets rid of the commit message textbox resizing.  This was
irritating to the user because they ended up clicking the "save"
button twice to save.

The code that cleared the default commit message was also moved out of
the two statically declared event handlers into two dynamically
attached handlers.  Instead of clearing that text on focus, the text
is selected so that typing will replace it.