Gardening: Make a new root dir in the repo called 'tools'.
This directory is going to contain all the tools that we write and use
for *developing*.

Start off by moving /boards/lib/bom-scripts to /tools/bom.

Create /tools/wrapper.  This will later contain a wrapper script that
can be used to invoke all the other scripts in the tools directory.

Create /tools/{pcb,fwman} too.