tools: Add a wrapper script that funnels SR devtool commands.
This is the Student Robotics development tool wrapper script.
It provides an interface to a number of different CLI tools used by
the SR project.

The idea is that this script searches for executable files in the
subdirectories of /tools/ in subversion and allows one to run them
like so:
 sr SCRIPT

Where SCRIPT is the name of an executable file with the path
/tools/*/SCRIPT in subversion.

This script *must* be checked out along with the rest of /tools.
See http://trac.srobo.org/wiki/DevScripts for more details.