roboide: Don't use the 'rev' property from file listings.
The rev property given by /filelist for each file/directory is
actually the revision in which the file was created.  I guess this
problem arose because the docs say:

"created_rev - pysvn.Revision - the revision of the last change"

which is rather self-contradicting, as the variable name "created_rev"
sounds like it's the version in which the file was created.  Turns out
the variable name is right...

This patch gets rid of the 'ide_rev' property from each item in the
file listing.  When one clicks on a file, the filelist will open up
the file using the revision number that it requested the list for.

As a side effect, this patch also cleans up referring to revisions.
The file list now uses "HEAD" to refer to HEAD, rather than 0.