pyenv: Clean up the motor interface.
Provide a "motor" object that can be used like this:
  motor[0] = 30
To set motor 0 to 30% speed.
motor[0] can also be read to find out the current speed.

This patch removes multiple implementations of the same I2C
transaction and condenses them into the same function for easier
maintenance.  Whilst doing this, I discovered that the routines that
read the motor power would return 0 in the event of failure.  They now
retry until there's no failure.