pyenv: Return information about the event that occurred in the event object.
robot.event now gets set to be an instance of the EventInfo class.
This takes a tree-like list of events that occurred and requests that
they each add properties to the EventInfo instance to present
information about the list.

The EventInfo class also overrides the == operator, so that if there
were multiple sources of events, the event appears to have come from
both sources, e.g.:

{{{
yield io.pin[1] == power.switch[0]

if event == io:
   print "Things"
}}}

The "== io" could have been changed here for "== power_switch" and the
code would function in the same way.