power fw: Keep the slug rail on from the start.
When the slug rail comes up on (some?) powerboards, the MSP430 ends up
getting reset.  This seems to be due to some power rail flailing.

Tom and I believe that this is because the switch-mode supplies can't
soft-start because of their lack of current-sensing resistors.

Tom didn't have this issue on his prototype powerboard at home, which
has current sense resistors correctly fitted.

This patch represents the last 4 hours of work.  Yay.