motor pcb: Add a footprint for a TVS diode.
This change doesn't change any of the existing topology.  It merely
adds an additional diode, which can optionally be fitted.

TVS diodes are used for voltage spike supression.  I believe that
these could be a source of problems on the motor controller, so I've
placed space for a 16 V bi-directional TVS diode across the motor
board's supply rail.  Farnell part I have in mind: 1470983 (I've
bought a couple for testing purposes).  These can be fitted onto the
boards after assembly if adding it works as intended.