motor-fw: Add support for feedback sensors and control routines.
This patch:
 * Adds a general interface to sensor drivers.
 * Adds a general interface to control routines.
 * Adds a control routine that just passes through: "unity".
 * Adds a dummy sensor: "null".
 * Uses a new function for setting motor power/direction.
   This function takes a signed integer.  If it's negative, then the
   motor is set to reverse.  This (hopefully) makes the control stuff
   more straight-forward.

It is possible to use the unity and null drivers in concert to
make the motor controller behave as it used to.  This is the current
configuration.

Please note that this patch doesn't wire any I2C routines up to the
new control stuff, so it's strictly only controllable through the
debugger right now... I2C is coming.