motor-fw: Rework i2c routines to allow commands to be switched in and out.
This patch (which is rather large, but splitting it further didn't
make sense) lets the sensors, controller and control loop to be
configured over I2C.  It does this by changing the i2c routines to
have a table that's split into pages of 16 commands.  When a
controller, sensor or control loop is initialised, its I2C commands
are integrated into the existing table.

The 'i2c_desc' code massively simplifies the configuration of
variables that can be set over i2c.