flashb: Write .data to the msp430!
Previously .data was not sent.  It should have been!  This patch is a
quick fix to this problem, and merges the .data section with the .text
section before transmission.  This means that more data is possibly
transferred to the msp430 than necessary.  However, this is shippable
and there is little time :-D